+++ CHANGELOG.md
@@ -8,7 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Added
-- Inbox system events 
+- Inbox achievement subject
 
 ## [0.2.1] - 2024-03-01